refactor(store): Wave 4, second slice -- split oauth out of store.ts
Second implementation-moving PR of Phase 2's Store split, after the
telemetry pilot (#955). Same delegation mechanism: OAuthStore gets its
own class per backend in store/slices/oauth.ts, and InMemoryStore/
FirestoreStore hold an oauthStore field with one-line forwards.

Why oauth second: field-ownership analysis found it structurally
clean -- zero fields shared with any other slice, only one same-slice
internal call (getOAuthGrantByRefreshTokenId from
rotateOAuthRefreshToken) -- with exactly one exception: five of
InMemoryStore's OAuth Maps are also touched by deleteAccountIdentity
(erasing a user's OAuth footprint). This PR proves the small-
documented-exception pattern the wave 4 plan settled on for that
case, rather than refactoring GDPR-erasure code into a hook/
coordinator pattern mid-phase.

- New apps/api/src/store/slices/oauth.ts: OAuthStore interface plus
  InMemoryOAuthStore and FirestoreOAuthStore, moved verbatim -- no
  logic changes. InMemoryOAuthStore's five Maps are deliberately not
  `private`: deleteAccountIdentity needs to reach across all five to
  purge an erased account's grants/tokens/codes, and inventing a
  hook/coordinator abstraction to preserve `private` here would be
  the exact refactor of compliance-sensitive code the wave 4 plan
  chose to defer. FirestoreStore's deleteAccountIdentity needed *no*
  change at all: it already reaches oauth collections through
  `this.db` by string literal, never through FirestoreOAuthStore's
  internals -- confirming the plan's finding that this asymmetry
  between the two backends is real, not a design choice.
- Extracted `stripUndefined` into a new shared
  apps/api/src/store/firestore-util.ts: it's used by four slices
  still living in store.ts (Identity, Access, Contribution) besides
  OAuth, so moving it into oauth.ts would have created a real
  circular module dependency (oauth.ts importing back from store.ts).
  Pulling it into its own file avoids that import cycle now and sets
  up cleanly for the slices still to come. store-firestore.test.ts's
  `describe('stripUndefined', ...)` block moves to
  store/firestore-util.test.ts alongside it, for the same reason
  store.test.ts's telemetry-retention tests moved in the prior PR.
- store.ts: 7369 -> 7129 lines (280 fewer; the interface, both
  implementations, and stripUndefined's definition all left).

Comment-prose: two lines I wrote fresh (documenting why the five Maps
aren't private, in both oauth.ts and store.ts's deleteAccountIdentity)
trimmed to compliant one-liners before resealing, so the reseal
covers only genuinely relocated prose -- same discipline established
on #954 and #955.

Verified erase-account.test.ts specifically exercises the changed
deleteAccountIdentity path (creates an OAuth grant + access token,
erases the account, asserts both are gone) and it passed unchanged.

Wave 4 plan: docs/phase2-store-split-execution-plan.md (private ops
repo). Next: a shared PR growing fakeFirestore()'s query support
(orderBy, ranges, in, cursors -- currently ==/!= only), which
unblocks ten more slices at once.
