+++ eslint-rules/comment-prose-baseline.json
@@ -275,6 +275,7 @@
     "apps/api/src/store.test.ts": 486,
     "apps/api/src/store.ts": 16323,
     "apps/api/src/store/fake-firestore.ts": 328,
+    "apps/api/src/store/firestore-util.ts": 163,
     "apps/api/src/store/records/access-tokens.ts": 204,
     "apps/api/src/store/records/agent-keys.ts": 169,
     "apps/api/src/store/records/build-log.ts": 492,
@@ -290,6 +291,7 @@
     "apps/api/src/store/records/submission.ts": 1562,
     "apps/api/src/store/records/telemetry.test.ts": 92,
     "apps/api/src/store/records/telemetry.ts": 1136,
+    "apps/api/src/store/slices/oauth.ts": 83,
     "apps/api/src/store/slices/telemetry.ts": 167,
     "apps/api/src/submission-status.test.ts": 35,
     "apps/api/src/submission-status.ts": 257,
